Fundamental now only logs the fundamental at times it's queried, and "interpolates" for the times between. Observations based off of time series objects now appropriately handle for this sparsity by making the same interpolation assumptions as TimeSeries. Note that this interpolation is essentially hard coded as it's basically the only logical way to interpolate that allows us to easily compute statistics in linear time with the number of samples, not the length of time.
The interpolation assumes that the value in unchanged after the last
observation, which is logical in the case of spread and median, and is somewhat
biased in terms of the fundamental. However, when we're concerned about
variance, no interpolation will accurately account, without actually
estimating the variance which would be very involved for likely little
gain.

The sampling for periodic measures is still accurate under the assumptions of
the interpolation, which will probably result in more accurate proxies than
using the point-by-point or 1-period measures, especially for values like
variance.

+++ hft-sim/src/entity/agent/strategy/FinalFundamentalEstimator.java
@@ -12,12 +12,12 @@ public class FinalFundamentalEstimator {
 
 	private final FundamentalValueView fundamental;
 	private final Timeline timeline;
-	private final int simulationLength;
+	private final long simulationLength;
 	private final double kappa;
-	private final int fundamentalMean;
+	private final double fundamentalMean;
 	
 	private FinalFundamentalEstimator(FundamentalValueView fundamental, Timeline timeline,
-			int simulationLength, double kappa, int fundamentalMean) {
+			long simulationLength, double kappa, double fundamentalMean) {
 		this.fundamental = fundamental;
 		this.timeline = timeline;
 		this.simulationLength = simulationLength;
@@ -26,12 +26,11 @@ public class FinalFundamentalEstimator {
 	}
 	
 	public static FinalFundamentalEstimator create(FundamentalValueView fundamental, Timeline timeline,
-			int simulationLength, double kappa, int fundamentalMean) {
+			long simulationLength, double kappa, double fundamentalMean) {
 		return new FinalFundamentalEstimator(fundamental, timeline, simulationLength, kappa, fundamentalMean);
 	}
 	
-	public static FinalFundamentalEstimator create(FundamentalValueView fundamental, Timeline timeline,
-			Props props) {
+	public static FinalFundamentalEstimator create(FundamentalValueView fundamental, Timeline timeline, Props props) {
 		return create(fundamental, timeline, props.get(SimLength.class), props.get(FundamentalKappa.class), props.get(FundamentalMean.class));
 	}
 
@@ -39,8 +38,7 @@ public class FinalFundamentalEstimator {
 	public Price getFundamentalEstimate() {
 		double stepsLeft = simulationLength - 1 - timeline.getCurrentTime().getInTicks() + fundamental.getLatency().getInTicks();
 		double kappaCompToPower = Math.pow(1 - kappa, stepsLeft);
-		return Price.of(fundamental.getValue().doubleValue() * kappaCompToPower 
-			+ fundamentalMean * (1 - kappaCompToPower));
+		return Price.of(fundamental.getValue().doubleValue() * kappaCompToPower + fundamentalMean * (1 - kappaCompToPower));
 	}
 	
 }
